refactor: reorganized nitrodigest packages
Created a cli package under Nitrodigest that contains all files needed for CLI.
