--- Projects/Nitrodigest/Cli/src/__init__.py
@@ -1,2 +1,0 @@
-"""nitrodigest CLI package"""
-__version__ = "0.1.2"
