+++ Projects/Nitrodigest/setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = nitrodigest-cli
-version = 0.1.2
+version = 0.1.3
 author = Marcin Kwiatkowski
 author_email = marcin@frodigo.com
 description = The privacy‑first, local‑LLM text‑summariser for developers.
@@ -15,8 +15,6 @@ package_dir =
     = src
 packages = find:
 python_requires = >=3.6
-py_modules =
-    main
 
 [options.packages.find]
 where = src
@@ -30,7 +28,7 @@ exclude =
 
 [options.entry_points]
 console_scripts =
-    nitrodigest = main:main
+    nitrodigest = cli:main
 
 [project]
 license = "MIT"
