+++ Projects/Nitrodigest/src/config.json.example
@@ -2,5 +2,5 @@
     "model": "mistral",
     "ollama_api_url": "http://localhost:11434",
     "timeout": 300,
-    "prompt_file": "summarizer/prompt_template.txt"
+    "prompt_file": "prompt_template.txt"
 }
