+++ Projects/Nitrodigest/src/prompt_template.txt
