+++ Projects/Nitrodigest/src/run-nitrodigest-cli.py
@@ -1,0 +1,4 @@
+from cli.main import main
+
+if __name__ == "__main__":
+    main()
